requirements.txt
streamlit
langchain-community
openai
python-dotenv
yfinance
sentence-transformers
scikit-learn
pinecone